Move SetDefaultEventuallyTimeout in the correct place
SetDefaultEventuallyTimeout was set after the TestCase.CheckDeployment
steps. This means that any CheckDeployment step could timeout if it used
an Eventually statement, and this is the case of nfs.

We probably did not hit this issue in our pipelines since we run many
testcases and not just one. If at least one testcase without an
Eventually step in the CheckDeployment would run, the test suite then
would SetDefaultEventuallyTimeout properly for the rest of the
testcases.