Embed fixtures to fix package import issue
Note from now on changes to the fixtures app will
require running `go generate`.

The problem with the current implementation is we expect when
this package is imported by another project using go modules,
that all the files are stored in the module cache. This does
not seem to be true for embedded modules, most notably
credhub-test-app. By embedding archived versions of the fixtures
we get around this issue. Simply embedding the directories
like `//go:embed credhub-test-app test_app` did not work
and returned a build time error like this:
   golang cannot embed directory $DIR: in different module

I don't really like this because it adds an additional step
and could easily be missed when making changes to the fixtures;
however, I tried a few other approaches and this one seemed the
cleanest. I would not be opposed to using a Makefile instead
of `go generate` because it would be more discoverable to
contributors.

A/C:
Given the testcases or runner package are imported into a
      different module
When some specific fixture needs to be used
Then the fixture is written to a provided temporary directory

I tested this with a smith environment and the acceptance tests
are not passing yet.